requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.10
numpy>=1.20.0
requests>=2.25.0
